homework for lesson 2
