+++ lesson2/substrate-node-template/Cargo.toml
@@ -4,6 +4,6 @@ panic = 'unwind'
 [workspace]
 members = [
     'node',
-    'pallets/template',
+    'pallets/poe',
     'runtime',
 ]
