+++ lesson2/substrate-node-template/runtime/src/lib.rs
@@ -40,8 +40,8 @@ pub use frame_support::{
 	},
 };
 
-/// Importing a template pallet
-pub use template;
+/// Importing the poe pallet
+pub use poe;
 
 /// An index to a block.
 pub type BlockNumber = u32;
@@ -252,8 +252,7 @@ impl sudo::Trait for Runtime {
 	type Call = Call;
 }
 
-/// Used for the module template in `./template.rs`
-impl template::Trait for Runtime {
+impl poe::Trait for Runtime {
 	type Event = Event;
 }
 
@@ -271,8 +270,7 @@ construct_runtime!(
 		Balances: balances::{Module, Call, Storage, Config<T>, Event<T>},
 		TransactionPayment: transaction_payment::{Module, Storage},
 		Sudo: sudo::{Module, Call, Config<T>, Storage, Event<T>},
-		// Used for the module template in `./template.rs`
-		TemplateModule: template::{Module, Call, Storage, Event<T>},
+		PoeModule: poe::{Module, Call, Storage, Event<T>},
 	}
 );
